backend/requirements.txt
Flask>=2.2.0
flask-cors>=3.0.10
python-dotenv>=0.21.0
SQLAlchemy>=1.4.0
psycopg2-binary>=2.9.0
numpy>=1.24.0
scipy>=1.10.0
Flask>=2.2.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
flask-cors>=3.0.10
